x10 performance on save ppm image by printing integers instead of floats

+++ renderer/shared/include/shared/vec3.cuh
@@ -285,7 +285,8 @@ struct rgb : public vec3
 
 inline std::ostream& operator<<(std::ostream& is, const rgb& t)
 {
-    is << t.e[0] * 255.99 << " " << t.e[1] * 255.99 << " " << t.e[2] * 255.99;
+    is << static_cast<unsigned int>(t.e[0] * 255.99) << " " << static_cast<unsigned int>(t.e[1] * 255.99)
+       << " " << static_cast<unsigned int>(t.e[2] * 255.99);
     return is;
 }
